requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==36.0.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==25.0
pycparser==2.22
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
celery==5.4.0
redis==5.0.7
